requirements.txt
Django==4.2.0
djangorestframework==3.14.0
Pillow==9.5.0
python-decouple==3.8
django-cors-headers==3.14.0
PyJWT==2.7.0
qrcode==7.4.2